log bugfixes

+++ src/log.h
@@ -17,6 +17,6 @@ typedef struct{
 esp_err_t log_init();
 void log_add_element_f(char *name, float *vals, size_t num_vals, int index);
 void log_update_vals_f(uint8_t index, float *vals, size_t num_vals);
-void log_add_element_i(char *name, float *vals, size_t num_vals, int index);
-void log_update_vals_i(uint8_t index, float *vals, size_t num_vals);
+void log_add_element_i(char *name, int *vals, size_t num_vals, int index);
+void log_update_vals_i(uint8_t index, int *vals, size_t num_vals);
 void log_output_task(void *arg);
